Address review comments
Merge test_wo_qnn_canonicalization.py into test_qnn_op_integration.py.
