+++ python/tvm/topi/hexagon/qnn/nn.py
@@ -22,8 +22,14 @@ import numpy as np
 
 import tvm
 from tvm import te, topi
-from ..utils import saturate, get_fixed_point_value
-from ...utils import get_const_tuple, get_const_int, get_const_float
+from ..utils import (
+    saturate,
+    is_scalar,
+    get_const_int_value,
+    get_const_float_value,
+    get_fixed_point_value,
+)
+from ...utils import get_const_tuple
 from ...nn.utils import get_pad_tuple
 from ...nn.pad import pad
 from ... import tag, nn
@@ -38,27 +44,6 @@ def clip_cast(val, dtype):
     return te.max(tvm.te.min(val, const_max), const_min).astype(dtype)
 
 
-# Return True if given expression is scalar constant value.
-def is_scalar(expr):
-    if isinstance(expr, te.Tensor):
-        return expr.ndim == 0 and (isinstance(expr.op.body[0], (tvm.tir.FloatImm, tvm.tir.IntImm)))
-    return isinstance(expr, (tvm.tir.FloatImm, tvm.tir.IntImm))
-
-
-def get_const_int_value(expr):
-    if isinstance(expr, te.Tensor):
-        assert isinstance(expr.op.body[0], tvm.tir.IntImm)
-        return expr.op.body[0].value
-    return get_const_int(expr)
-
-
-def get_const_float_value(expr):
-    if isinstance(expr, te.Tensor):
-        assert isinstance(expr.op.body[0], tvm.tir.FloatImm)
-        return expr.op.body[0].value
-    return get_const_float(expr)
-
-
 def get_qnn_param(param, indices, axis):
     # Account scalar and 1D quantization parameters:
     if len(param.shape) == 0:
@@ -68,11 +53,7 @@ def get_qnn_param(param, indices, axis):
     return param[param_idx]
 
 
-def subtract_zero_point(
-    tensor: te.Tensor,
-    zero_point: Union[te.Tensor, tvm.tir.IntImm],
-    name: str,
-):
+def subtract_zero_point(tensor: te.Tensor, zero_point: Union[te.Tensor, tvm.tir.IntImm], name: str):
     """
     Subtract zero point from given tensor. If zero point is scalar constant and is equal to 0, then
     it can be optimized and return tensor as it is.
@@ -590,10 +571,7 @@ def qnn_conv2d(  # Conv2d inputs
         oshape,
         lambda n, oc, oh, ow: te.sum(
             data_pad[
-                n,
-                ic,
-                oh * height_stride + kh * dilation_h,
-                ow * width_stride + kw * dilation_w,
+                n, ic, oh * height_stride + kh * dilation_h, ow * width_stride + kw * dilation_w
             ].astype("int32")
             * weight[oc, ic, kh, kw].astype("int32"),
             axis=[ic, kh, kw],
@@ -777,10 +755,7 @@ def qnn_depthwise_conv2d(  # Conv2d inputs
         oshape,
         lambda n, oc, oh, ow: te.sum(
             data_pad[
-                n,
-                oc,
-                oh * height_stride + kh * dilation_h,
-                ow * width_stride + kw * dilation_w,
+                n, oc, oh * height_stride + kh * dilation_h, ow * width_stride + kw * dilation_w
             ].astype("int32")
             * te.subtract(weight[oc, 0, kh, kw], kernel_zero_point).astype("int32"),
             axis=[kh, kw],
